+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+-   Update `set_name` to not add legacy Mac name records to fonts that don't have them (Mac records are updated only if already present).
 ## [0.28.0](https://github.com/fabiocaccamo/python-fontbro/releases/tag/0.28.0) - 2026-09-11
 -   Add `get_supported_languages` and `get_supported_writing_systems` methods based on the Google Fonts languages dataset.
 -   Add `get_embedding_permissions` and `set_embedding_permissions` methods for OS/2 `fsType` embedding flags.